+++ src/main/java/com/alkemy/ong/controllers/AuthController.java
@@ -10,8 +10,8 @@ import com.alkemy.ong.services.SendGridService;
 import com.alkemy.ong.pojos.output.UserProfileDTO;
 import com.alkemy.ong.repositories.UserRepository;
 import com.alkemy.ong.services.RoleService;
-import com.alkemy.ong.services.UserDetailsServices;
 import com.alkemy.ong.services.UserService;
+import com.alkemy.ong.services.impl.UserDetailsServices;
 import com.alkemy.ong.util.JwtUtil;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.config.Configuration;
@@ -22,7 +22,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.validation.BindingResult;
