+++ src/main/java/com/alkemy/ong/controllers/ContactController.java
@@ -3,7 +3,6 @@ package com.alkemy.ong.controllers;
 import com.alkemy.ong.dtos.requests.ContactPostDTO;
 import com.alkemy.ong.dtos.responses.ContactListDTO;
 import com.alkemy.ong.entities.Contact;
-import com.alkemy.ong.entities.User;
 import com.alkemy.ong.services.ContactService;
 import com.alkemy.ong.services.SendGridService;
 import com.alkemy.ong.services.UserService;
@@ -14,7 +13,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
@@ -39,7 +37,7 @@ public class ContactController {
         Contact contactCreated;
        // try {
             Contact contactToCreate = contactPostDto.toContact();
-            contactCreated = contactService.createContact(contactToCreate);
+            contactCreated = contactService.create(contactToCreate);
 
             //Contact Mail Sending
             httpResponse.addHeader("User-Mail-Sent", String.valueOf(sendGridService.contactMessage(contactPostDto.getName(), contactPostDto.getEmail())));
@@ -58,6 +56,6 @@ public class ContactController {
     @GetMapping
     public List<ContactListDTO> getAll() {
         //Falta validación como administrador
-        return contactService.findAllContacts();
+        return contactService.findAll();
     }
 }
