+++ src/main/java/com/alkemy/ong/util/JwtFilter.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.util;
 
-import com.alkemy.ong.services.UserDetailsServices;
+import com.alkemy.ong.services.impl.UserDetailsServices;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
